Refactor setUp method into test utils
