+++ test/ComposableStable.integration.spec.ts
@@ -8,19 +8,13 @@ import {
 } from '@balancer-labs/typechain';
 import { vaultAddr } from './testScripts/constants';
 import { SubgraphPoolBase, SwapTypes, SOR, bnum } from '../src';
-import {
-    Network,
-    MULTIADDR,
-    SOR_CONFIG,
-    ADDRESSES,
-} from './testScripts/constants';
-import { OnChainPoolDataService } from './lib/onchainData';
-import { TokenPriceService } from '../src';
+import { Network, ADDRESSES } from './testScripts/constants';
 import { AddressZero } from '@ethersproject/constants';
 import { BigNumber, parseFixed } from '@ethersproject/bignumber';
 import { expect } from 'chai';
 import { closeTo } from './lib/testHelpers';
 import { ComposableStablePool } from '../src/pools/composableStable/composableStablePool';
+import { setUp } from './testScripts/utils';
 
 dotenv.config();
 
@@ -28,6 +22,7 @@ let sor: SOR;
 const networkId = Network.MAINNET;
 const { ALCHEMY_URL: jsonRpcUrl } = process.env;
 const rpcUrl = 'http://127.0.0.1:8545';
+const blockNumber = 16447247;
 const provider = new JsonRpcProvider(rpcUrl, networkId);
 const vault = Vault__factory.connect(vaultAddr, provider);
 const bbausdt = ADDRESSES[networkId].bbausdt2.address;
@@ -39,7 +34,6 @@ const funds = {
     fromInternalBalance: false,
     toInternalBalance: false,
 };
-let subgraphPoolDataService: OnChainPoolDataService;
 
 // bbausd
 const testPool: SubgraphPoolBase = {
@@ -144,39 +138,6 @@ const testPool1: SubgraphPoolBase = {
     totalWeight: '0',
 };
 
-// Setup SOR with data services
-function setUp(networkId: Network, provider: JsonRpcProvider): SOR {
-    // The SOR needs to fetch pool data from an external source. This provider fetches from Subgraph and onchain calls.
-    subgraphPoolDataService = new OnChainPoolDataService({
-        vaultAddress: vaultAddr,
-        multiAddress: MULTIADDR[networkId],
-        provider,
-        pools: [testPool, testPool1],
-    });
-
-    class CoingeckoTokenPriceService implements TokenPriceService {
-        constructor(private readonly chainId: number) {}
-        async getNativeAssetPriceInToken(
-            // eslint-disable-next-line @typescript-eslint/no-unused-vars
-            tokenAddress: string
-        ): Promise<string> {
-            return '0';
-        }
-    }
-
-    // Use coingecko to fetch token price information. Used to calculate cost of additonal swaps/hops.
-    const coingeckoTokenPriceService = new CoingeckoTokenPriceService(
-        networkId
-    );
-
-    return new SOR(
-        provider,
-        SOR_CONFIG[networkId],
-        subgraphPoolDataService,
-        coingeckoTokenPriceService
-    );
-}
-
 export async function queryJoin(
     network: number,
     poolId: string,
@@ -248,24 +209,20 @@ export async function querySingleTokenExit(
 }
 
 describe('ComposableStable', () => {
+    // Setup chain
+    before(async function () {
+        this.timeout(20000);
+
+        sor = await setUp(
+            networkId,
+            provider,
+            [testPool, testPool1],
+            jsonRpcUrl as string,
+            blockNumber
+        );
+        await sor.fetchPools();
+    });
     context('test swaps vs queryBatchSwap', () => {
-        // Setup chain
-        before(async function () {
-            this.timeout(20000);
-
-            await provider.send('hardhat_reset', [
-                {
-                    forking: {
-                        jsonRpcUrl,
-                        blockNumber: 16447247,
-                    },
-                },
-            ]);
-
-            sor = setUp(networkId, provider);
-            await sor.fetchPools();
-        });
-
         context('ExactIn', () => {
             it('token>token', async () => {
                 const swapType = SwapTypes.SwapExactIn;
@@ -404,22 +361,6 @@ describe('ComposableStable', () => {
         });
     });
     context('test joins vs queryJoin', () => {
-        // Setup chain
-        before(async function () {
-            this.timeout(20000);
-
-            await provider.send('hardhat_reset', [
-                {
-                    forking: {
-                        jsonRpcUrl,
-                        blockNumber: 16447247,
-                    },
-                },
-            ]);
-
-            sor = setUp(networkId, provider);
-            await sor.fetchPools();
-        });
         context('Joins', () => {
             it('Join with many tokens', async () => {
                 const bptIndex = 2;
@@ -432,7 +373,7 @@ describe('ComposableStable', () => {
                 const amountsWithOutBpt = [...amountsInWithBpt];
                 amountsWithOutBpt.splice(bptIndex, 1);
 
-                const pools = await subgraphPoolDataService.getPools();
+                const pools = sor.getPools();
                 const pool = ComposableStablePool.fromPool(pools[0]);
                 const bptCalculated =
                     pool._calcBptOutGivenExactTokensIn(amountsWithOutBpt);
@@ -462,7 +403,7 @@ describe('ComposableStable', () => {
                 const amountsWithOutBpt = [...amountsInWithBpt];
                 amountsWithOutBpt.splice(bptIndex, 1);
 
-                const pools = await subgraphPoolDataService.getPools();
+                const pools = sor.getPools();
                 const pool = ComposableStablePool.fromPool(pools[0]);
                 const bptCalculated =
                     pool._calcBptOutGivenExactTokensIn(amountsWithOutBpt);
@@ -484,28 +425,12 @@ describe('ComposableStable', () => {
         });
     });
     context('test exits vs queryExit', () => {
-        // Setup chain
-        before(async function () {
-            this.timeout(20000);
-
-            await provider.send('hardhat_reset', [
-                {
-                    forking: {
-                        jsonRpcUrl,
-                        blockNumber: 16447247,
-                    },
-                },
-            ]);
-
-            sor = setUp(networkId, provider);
-            await sor.fetchPools();
-        });
         context('Exits', () => {
             it('BPT>token', async () => {
                 const tokenIndex = 0;
                 const bptInHuman = '977.234';
                 const bptInEvm = parseFixed(bptInHuman, 18);
-                const pools = await subgraphPoolDataService.getPools();
+                const pools = sor.getPools();
                 const pool = ComposableStablePool.fromPool(pools[0]);
                 const pairData = pool.parsePoolPairData(
                     testPool.address,
@@ -539,7 +464,7 @@ describe('ComposableStable', () => {
             //     const tokenIndex = 2; // usdc
             //     const bptInHuman = '10.10';
             //     const bptInEvm = parseFixed(bptInHuman, 18);
-            //     const pools = await subgraphPoolDataService.getPools();
+            //     const pools = sor.getPools();
             //     const pool = ComposableStablePool.fromPool(pools[1]);
             //     const pairData = pool.parsePoolPairData(
             //         pool.address,
@@ -572,7 +497,7 @@ describe('ComposableStable', () => {
             // ComposableStable V1 does not have this functionality but V2 does
             // it('BPT>tokens', async () => {
             //     const bptIn = parseFixed('77', 18);
-            //     const pools = await subgraphPoolDataService.getPools();
+            //     const pools = sor.getPools();
             //     const pool = ComposableStablePool.fromPool(pools[0]);
             //     const amountOut = pool._calcTokensOutGivenExactBptIn(bptIn);
             //     console.log(amountOut.toString());
